% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dolt-remote.R
\name{dolt_creds}
\alias{dolt_creds}
\alias{dolt_creds_new}
\title{Credentials for authenticated remotes}
\usage{
dolt_creds(con, action = "export", id = NULL, path = NULL)

dolt_creds_new(con)
}
\arguments{
\item{con}{a `DoltliteConnection`.}

\item{action}{`"export"`.}

\item{id}{a credential id.}

\item{path}{a directory to export the public JWK into. Omit to have the
JWK returned instead.}
}
\value{
A string.
}
\description{
Credentials for authenticated remotes
}
