% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dolt-commit.R
\name{dolt_gc}
\alias{dolt_gc}
\title{Garbage-collect unreachable chunks}
\usage{
dolt_gc(con)
}
\arguments{
\item{con}{a `DoltliteConnection`.}
}
\value{
A summary string, e.g. `"12 chunks removed, 45 chunks kept"`.
}
\description{
Garbage-collect unreachable chunks
}
