% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dolt-commit.R
\name{dolt_hashof}
\alias{dolt_hashof}
\alias{dolt_hashof_table}
\alias{dolt_hashof_db}
\title{Content-addressed hashes}
\usage{
dolt_hashof(con, ref = "HEAD")

dolt_hashof_table(con, table, ref = NULL)

dolt_hashof_db(con, ref = NULL)
}
\arguments{
\item{con}{a `DoltliteConnection`.}

\item{ref}{a branch, tag, commit hash, `HEAD`, `HEAD~N` or `HEAD^N`.}

\item{table}{a table name.}
}
\value{
A 40-character lowercase hex hash.
}
\description{
Content-addressed hashes
}
