% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dolt-branch.R
\name{dolt_merge_base}
\alias{dolt_merge_base}
\title{Merge base of two commits}
\usage{
dolt_merge_base(con, a, b)
}
\arguments{
\item{con}{a `DoltliteConnection`.}

\item{a, b}{commit hashes or refs.}
}
\value{
The merge-base commit hash.
}
\description{
Merge base of two commits
}
