% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dolt-remote.R
\name{dolt_push}
\alias{dolt_push}
\alias{dolt_fetch}
\title{Exchange commits with a remote}
\usage{
dolt_push(con, remote = "origin", branch = NULL)

dolt_fetch(con, remote = "origin", branch = NULL)
}
\arguments{
\item{con}{a `DoltliteConnection`.}

\item{remote}{the remote name, `"origin"` by default.}

\item{branch}{the branch to transfer. Omit for the remote's default
behaviour.}
}
\value{
The result string.
}
\description{
Exchange commits with a remote
}
