% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dolt-remote.R
\name{dolt_remote}
\alias{dolt_remote}
\alias{dolt_remotes}
\title{Manage remotes}
\usage{
dolt_remote(con, action = NULL, name = NULL, url = NULL)

dolt_remotes(con)
}
\arguments{
\item{con}{a `DoltliteConnection`.}

\item{action}{`"add"` or `"remove"`. Omit to list remotes.}

\item{name}{the remote name, e.g. `"origin"`.}

\item{url}{a `file://` or `http(s)://` URL. The HTTP form includes the
database name, e.g. `http://host:8080/mydb.db`.}
}
\value{
A data frame when listing, otherwise invisibly `TRUE`.
}
\description{
Manage remotes
}
\examples{
con <- DBI::dbConnect(doltliter::Doltlite(), tempfile())
dolt_remote(con)
DBI::dbDisconnect(con)
}
