% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dolt-commit.R
\name{dolt_revert}
\alias{dolt_revert}
\title{Revert a commit}
\usage{
dolt_revert(con, ref)
}
\arguments{
\item{con}{a `DoltliteConnection`.}

\item{ref}{a commit hash or ref to revert.}
}
\value{
The result string: a new commit hash, or a message reporting
  conflicts.
}
\description{
Creates a new commit applying the inverse of `ref`. The initial commit
cannot be reverted.
}
