% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dolt-commit.R
\name{dolt_verify_constraints}
\alias{dolt_verify_constraints}
\title{Re-check constraints}
\usage{
dolt_verify_constraints(con, tables = NULL, all = FALSE, output_only = FALSE)
}
\arguments{
\item{con}{a `DoltliteConnection`.}

\item{tables}{optional table names; omit to check all.}

\item{all}{pass `--all`.}

\item{output_only}{pass `--output-only`.}
}
\value{
The function's return value.
}
\description{
Re-check constraints
}
