% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{doltlite_engine}
\alias{doltlite_engine}
\alias{doltlite_version}
\title{Library information}
\usage{
doltlite_engine()

doltlite_version()
}
\value{
`doltlite_engine()` returns the storage engine name, `"prolly"` for
  a DoltLite-format database. `doltlite_version()` returns the bundled
  SQLite version.
}
\description{
Library information
}
