% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-classes.R
\name{DOLTLITE_RO}
\alias{DOLTLITE_RO}
\alias{DOLTLITE_RW}
\alias{DOLTLITE_RWC}
\title{Database open flags}
\usage{
DOLTLITE_RO

DOLTLITE_RW

DOLTLITE_RWC
}
\description{
Passed as `flags` to [DBI::dbConnect()]. These are SQLite's
`SQLITE_OPEN_*` values.
}
