% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-classes.R, R/Doltlite-connection.R
\docType{class}
\name{DoltliteConnection-class}
\alias{DoltliteConnection-class}
\alias{dbIsValid,DoltliteConnection-method}
\alias{dbDisconnect,DoltliteConnection-method}
\alias{show,DoltliteConnection-method}
\alias{dbGetInfo,DoltliteConnection-method}
\alias{dbSendQuery,DoltliteConnection,character-method}
\alias{dbSendStatement,DoltliteConnection,character-method}
\alias{dbGetQuery,DoltliteConnection,character-method}
\alias{dbExecute,DoltliteConnection,character-method}
\alias{dbListTables,DoltliteConnection-method}
\alias{dbExistsTable,DoltliteConnection,character-method}
\alias{dbListFields,DoltliteConnection,character-method}
\alias{dbRemoveTable,DoltliteConnection,character-method}
\alias{dbReadTable,DoltliteConnection,character-method}
\alias{dbQuoteIdentifier,DoltliteConnection,character-method}
\alias{dbQuoteIdentifier,DoltliteConnection,SQL-method}
\alias{dbQuoteString,DoltliteConnection,character-method}
\alias{dbQuoteString,DoltliteConnection,SQL-method}
\alias{dbExistsTable,DoltliteConnection,SQL-method}
\alias{dbListFields,DoltliteConnection,SQL-method}
\alias{dbReadTable,DoltliteConnection,SQL-method}
\alias{dbRemoveTable,DoltliteConnection,SQL-method}
\title{DoltLite connection class}
\usage{
\S4method{dbIsValid}{DoltliteConnection}(dbObj, ...)

\S4method{dbDisconnect}{DoltliteConnection}(conn, ...)

\S4method{show}{DoltliteConnection}(object)

\S4method{dbGetInfo}{DoltliteConnection}(dbObj, ...)

\S4method{dbSendQuery}{DoltliteConnection,character}(conn, statement, params = NULL, ..., immediate = FALSE)

\S4method{dbSendStatement}{DoltliteConnection,character}(conn, statement, params = NULL, ..., immediate = FALSE)

\S4method{dbGetQuery}{DoltliteConnection,character}(conn, statement, params = NULL, ..., n = -1, immediate = FALSE)

\S4method{dbExecute}{DoltliteConnection,character}(conn, statement, params = NULL, ..., immediate = FALSE)

\S4method{dbListTables}{DoltliteConnection}(conn, ...)

\S4method{dbExistsTable}{DoltliteConnection,character}(conn, name, ...)

\S4method{dbListFields}{DoltliteConnection,character}(conn, name, ...)

\S4method{dbRemoveTable}{DoltliteConnection,character}(conn, name, ..., temporary = FALSE, fail_if_missing = TRUE)

\S4method{dbReadTable}{DoltliteConnection,character}(conn, name, ..., row.names = FALSE, check.names = TRUE)

\S4method{dbQuoteIdentifier}{DoltliteConnection,character}(conn, x, ...)

\S4method{dbQuoteIdentifier}{DoltliteConnection,SQL}(conn, x, ...)

\S4method{dbQuoteString}{DoltliteConnection,character}(conn, x, ...)

\S4method{dbQuoteString}{DoltliteConnection,SQL}(conn, x, ...)

\S4method{dbExistsTable}{DoltliteConnection,SQL}(conn, name, ...)

\S4method{dbListFields}{DoltliteConnection,SQL}(conn, name, ...)

\S4method{dbReadTable}{DoltliteConnection,SQL}(conn, name, ...)

\S4method{dbRemoveTable}{DoltliteConnection,SQL}(conn, name, ...)
}
\arguments{
\item{dbObj, drv, conn}{a `DoltliteConnection`.}

\item{...}{passed on to methods.}

\item{statement}{an SQL string.}

\item{params}{optional query parameters, as a list.}

\item{immediate}{whether to execute the statement immediately.}

\item{name}{a table name.}

\item{x}{a string or identifier to quote.}
}
\description{
DoltLite connection class
}
\section{Slots}{

\describe{
\item{\code{ptr}}{external pointer to the C connection object.}

\item{\code{dbname}}{the database path, without any branch suffix.}

\item{\code{branch}}{the branch requested at connect time, or `NA`.}

\item{\code{bigint}}{how out-of-range 64-bit integers are returned.}

\item{\code{state}}{an environment holding mutable per-connection state, chiefly
whether a transaction is open. S4 slots are copy-on-modify, so mutable
bookkeeping has to live in a reference object.}
}}

\keyword{internal}
