% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-classes.R, R/Doltlite-result.R
\docType{class}
\name{DoltliteResult-class}
\alias{DoltliteResult-class}
\alias{dbIsValid,DoltliteResult-method}
\alias{dbClearResult,DoltliteResult-method}
\alias{dbFetch,DoltliteResult-method}
\alias{dbHasCompleted,DoltliteResult-method}
\alias{dbGetRowCount,DoltliteResult-method}
\alias{dbGetRowsAffected,DoltliteResult-method}
\alias{dbGetStatement,DoltliteResult-method}
\alias{dbColumnInfo,DoltliteResult-method}
\alias{dbBind,DoltliteResult-method}
\alias{show,DoltliteResult-method}
\title{DoltLite result class}
\usage{
\S4method{dbIsValid}{DoltliteResult}(dbObj, ...)

\S4method{dbClearResult}{DoltliteResult}(res, ...)

\S4method{dbFetch}{DoltliteResult}(res, n = -1, ...)

\S4method{dbHasCompleted}{DoltliteResult}(res, ...)

\S4method{dbGetRowCount}{DoltliteResult}(res, ...)

\S4method{dbGetRowsAffected}{DoltliteResult}(res, ...)

\S4method{dbGetStatement}{DoltliteResult}(res, ...)

\S4method{dbColumnInfo}{DoltliteResult}(res, ...)

\S4method{dbBind}{DoltliteResult}(res, params, ...)

\S4method{show}{DoltliteResult}(object)
}
\arguments{
\item{...}{unused.}

\item{res, dbObj}{a `DoltliteResult`.}

\item{n}{number of rows to fetch; `-1` (the default) means all remaining.}

\item{params}{a list of parameter values.}
}
\description{
DoltLite result class
}
\section{Slots}{

\describe{
\item{\code{ptr}}{external pointer to the C result object.}

\item{\code{conn}}{the connection that produced this result.}

\item{\code{sql}}{the statement text.}

\item{\code{is_statement}}{whether the statement returns no columns.}
}}

\keyword{internal}
