% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{reexports}
\alias{reexports}
\alias{Id}
\alias{SQL}
\alias{dbCanConnect}
\alias{dbIsReadOnly}
\alias{dbListObjects}
\alias{dbQuoteLiteral}
\alias{dbUnquoteIdentifier}
\alias{dbWithTransaction}
\title{Objects re-exported from DBI}
\arguments{
\item{...}{passed on to the DBI function.}
}
\value{
As documented by DBI.
}
\description{
These are DBI's own functions, re-exported so that they are available after
`library(doltliter)` without attaching DBI as well. See the linked DBI
documentation for each.
}
\seealso{
[DBI::Id()], [DBI::SQL()], [DBI::dbCanConnect()],
  [DBI::dbIsReadOnly()], [DBI::dbListObjects()], [DBI::dbQuoteLiteral()],
  [DBI::dbUnquoteIdentifier()], [DBI::dbWithTransaction()]
}
\keyword{internal}
